+++ inc/shootem_math.h
@@ -3,6 +3,6 @@
 class ShootEmMath {
  public:
   static float RandomNumber(float min, float max);
-  static void CalculateSlope(int x1, int y1, int x2, int y2, float& dx,
+  static void CalculateSlope(float x1, float y1, float x2, float y2, float& dx,
                              float& dy);
 };
